requirements.txt
alembic==1.12.1
amqp==5.2.0
async-timeout==4.0.3
billiard==4.2.0
blinker==1.7.0
boto3==1.28.85
botocore==1.31.85
celery==5.3.5
Cerberus==1.3.5
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Flask==3.0.0
Flask-Cors==4.0.0
flask-marshmallow==0.15.0
Flask-Migrate==4.0.5
Flask-Script==2.0.6
flask-serialize==2.1.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.3.3
Mako==1.3.0
MarkupSafe==2.1.3
marshmallow==3.20.1
packaging==23.2
permissive-dict==1.0.4
prompt-toolkit==3.0.41
python-dateutil==2.8.2
python-decouple==3.8
redis==5.0.1
s3transfer==0.7.0
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
uWSGI==2.0.23
vine==5.1.0
wcwidth==0.2.10
Werkzeug==3.0.1
